refactor doc-parser into modules
https://coveord.atlassian.net/browse/KIT-402
