+++ packages/headless/doc-parser/src/entity.ts
@@ -1,0 +1,11 @@
+export interface Entity {
+  name: string;
+  type: string;
+  isOptional: boolean;
+  desc: string;
+  // notes?: []
+}
+
+export interface ObjEntity extends Entity {
+  members: Entity[];
+}
